started on node process
