+++ shared/src/Unison/Remote.hs
@@ -5,13 +5,62 @@
 
 module Unison.Remote where
 
-import qualified Unison.Hashable as Hashable
-import Unison.Hashable (Hashable, Hashable1)
-import GHC.Generics
+import Data.Aeson (ToJSON(..),FromJSON(..))
+import Data.ByteString (ByteString)
 import Data.Text (Text)
-import Data.Aeson (ToJSON,FromJSON)
-import qualified Unison.Hashable as H
+import Data.Text.Encoding (decodeUtf8, encodeUtf8)
+import GHC.Generics
+import Unison.Hashable (Hashable, Hashable1)
+import qualified Data.ByteString.Base64.URL as Base64
 import qualified Data.Text as Text
+import qualified Unison.Hashable as H
+import qualified Unison.Hashable as Hashable
+
+{-
+Representation of the Unison distributed programming API.
+
+    data Node
+    data Local a
+
+    data Remote a
+    at : Node -> a -> Remote a
+    instance Monad Remote
+
+    instance Monad Local
+    fork : Remote a -> Remote ()
+    local : Local a -> Remote a
+
+    root : Node -> Channel Packet
+    packet : Remote a -> Packet
+
+`Local` is roughly `IO`, the type of local effects, but has
+a few additional functions:
+
+    channel : Local (Channel a)
+    send : a -> Channel a -> Local ()
+    -- | Registers a callback in a map as a weak ref that sets an MVar
+    -- if weak ref becomes garbage, remove from the map
+    receiveAsync : Channel a -> Local (Local a)
+    -- Can be done a bit more efficiently perhaps
+    recieve : Channel a -> Local a
+    awaitAsync : Remote a -> Local (Local a)
+    await : Remote a -> Local a
+
+For the implementation, a remote computation consists of a step, which evaluates
+locally or transfers control to another node, or a step along with a continuation
+when the step's result is available. Conceptually represented by the following Haskell type:
+
+    data Remote r
+      = Step r
+      | forall x . Bind (Step x) (x -> Remote r)
+
+    data Step r = Local r | At Node r
+
+Since this data type would not be serializable as a Haskell
+data type (would require serializing arbitrary functions), both
+`Local` and the `x -> Remote r` are represented as _Unison_
+terms, giving the type:
+-}
 
 -- `t` will be a Unison term, generally
 data Remote t = Step (Step t) | Bind (Step t) t deriving (Generic,Generic1,Show,Eq,Foldable,Functor,Traversable)
@@ -98,24 +147,27 @@ we use some of the effects in `Local` to write the final step to
 a channel from which we `receive`.
 -}
 
-newtype Base64 = Base64 Text deriving (Eq,Ord,Generic,Show)
-instance ToJSON Base64
-instance FromJSON Base64
-instance Hashable Base64 where
-  tokens (Base64 txt) = [Hashable.Text txt]
-
-
 -- | A node is a host and a public key. For instance: `Node "unisonweb.org" key`
-data Node = Node { host :: Text, publicKey :: Base64 } deriving (Eq,Ord,Generic)
-instance ToJSON Node
-instance FromJSON Node
+data Node = Node { host :: Text, publicKey :: ByteString } deriving (Eq,Ord,Generic)
+
+instance ToJSON Node where toJSON (Node host key) = toJSON (host, decodeUtf8 (Base64.encode key))
+instance FromJSON Node where
+  parseJSON v = do
+    (host,key) <- parseJSON v
+    either fail (pure . Node host) (Base64.decode (encodeUtf8 key))
+
 instance Hashable Node where
-  tokens (Node host (Base64 key)) = [Hashable.Text host, Hashable.Text key]
+  tokens (Node host key) = [Hashable.Text host, Hashable.Bytes key]
 
 instance Show Node where
-  show (Node host (Base64 key)) = "http://" ++ Text.unpack host ++ "/" ++ Text.unpack key
+  show (Node host key) = "http://" ++ Text.unpack host ++ "/" ++ Text.unpack (decodeUtf8 (Base64.encode key))
 
-newtype Channel = Channel Base64 deriving (Eq,Ord,Generic,Show)
-instance ToJSON Channel
-instance FromJSON Channel
+newtype Channel = Channel ByteString deriving (Eq,Ord,Generic,Show)
+instance ToJSON Channel where toJSON (Channel c) = toJSON (decodeUtf8 (Base64.encode c))
+
+instance FromJSON Channel where
+  parseJSON v = do
+    txt <- parseJSON v
+    either fail (pure . Channel) (Base64.decode (encodeUtf8 txt))
+
 instance Hashable Channel where tokens (Channel c) = Hashable.tokens c
